[model] CreatePassengerNameRecord api update
https://releasenotes.developer.sabre.com/2019/06/06/create-passenger-name-record-v2-2-0-on-june-6-2019-2/

What's New
Added ability to create reservations for GDS and aggregator hotel chains at …/CreatePassengerNameRecordRQ/HotelBook.
Updated internal API EnhancedAirBookRQ to version 3.10.0.
Updated internal API PassengerDetailsRQ to version 3.4.0.
Modified RQ schema to allow for flight segment booking being optional (this allows customer to create PNRs consisting of only Miscellaneous segments)
Business Value
Starting with version 2.2.0, the API was enhanced with the ability to create reservations for GDS and aggregator hotel chains. The solution integrates multiple sources of content including the Sabre global distribution system (GDS) and non-GDS sources such as re-distributors, agency-sourced hotel content and aggregators including HotelBeds, GTA, Expedia Partner Solutions and Booking.com.

+++ src/Model/CreatePassengerNameRecord/Normalizer/CreatePassengerNameRecordRequestCreatePassengerNameRecordRQAirTaxItineraryInfosItineraryInfoItemReservationItemsNormalizer.php
@@ -10,7 +10,6 @@ declare(strict_types=1);
 
 namespace Ammonkc\SabreApi\Model\CreatePassengerNameRecord\Normalizer;
 
-use Symfony\Component\Serializer\Exception\InvalidArgumentException;
 use Symfony\Component\Serializer\Normalizer\DenormalizerAwareInterface;
 use Symfony\Component\Serializer\Normalizer\DenormalizerAwareTrait;
 use Symfony\Component\Serializer\Normalizer\DenormalizerInterface;
@@ -36,10 +35,10 @@ class CreatePassengerNameRecordRequestCreatePassengerNameRecordRQAirTaxItinerary
     public function denormalize($data, $class, $format = null, array $context = [])
     {
         if (!is_object($data)) {
-            throw new InvalidArgumentException();
+            return null;
         }
         $object = new \Ammonkc\SabreApi\Model\CreatePassengerNameRecord\CreatePassengerNameRecordRequestCreatePassengerNameRecordRQAirTaxItineraryInfosItineraryInfoItemReservationItems();
-        if (property_exists($data, 'Item')) {
+        if (property_exists($data, 'Item') && $data->{'Item'} !== null) {
             $object->setItem($this->denormalizer->denormalize($data->{'Item'}, 'Ammonkc\\SabreApi\\Model\\CreatePassengerNameRecord\\CreatePassengerNameRecordRequestCreatePassengerNameRecordRQAirTaxItineraryInfosItineraryInfoItemReservationItemsItem', 'json', $context));
         }
